Enforce a well-defined order of group application
Kinda dirty hackjob time!

This commit changes the internal representation of the bot permissions groups to
a list instead of a set.  This means that bot group membership will be tested in
a well-defined and predictable order, specifically the order in which the groups
are listed in the permissions.ini file, instead of whatever arbitrary order the
interpreter feels like using.

This avoids a situation of a user having two roles with different bot group
memberships, and getting permissions applied inconsistently or incorrectly as a
result.