Be more aggressive about self-destructing for conn fails
Be more aggressive about self-destructing if a problem is preventing us from
connecting to the server.  We do this by checking what kind of OS we're on and
sending an appropriate signal back to ourselves if appropriate, otherwise using
os._exit() to kill our own process indiscriminately.

While this prevents the bot from restarting its own event loop cleanly if its
connection is lost, given that I've seen it hang multiple times when it's
supposed to be restarting itself, I feel this is a better state to be in since
it allows for systemd or any other wrapper script to restart the bot from
scratch.

But yes, this is definitely a hacky solution.  It would be better to go through
and rework all the async stuff to ensure that exceptions are being handled
correctly and such.  One day maybe.