Bump discord.py version
Bump the version of the Discord library we use.  Bots running this software
have developed issues of being unabel to set their status message ("Playing...")
and I suspect this is because Discord is in the process of deprecating older API
versions.  In any case, bumping the library version fixes those issues.

In addition, a couple minor fixes are contained within.